rename competitions components into leaderboards components
