+++ frontend/src/components/leaderboard-columns.tsx
