+++ frontend/src/components/leaderboard-data-table.tsx
